+++ src/main/java/mods/belgabor/amtweaker/mods/amt/AMT.java
@@ -2,6 +2,7 @@ package mods.belgabor.amtweaker.mods.amt;
 
 import minetweaker.MineTweakerAPI;
 import mods.belgabor.amtweaker.mods.amt.handlers.Evaporator;
+import mods.belgabor.amtweaker.mods.amt.handlers.IceMaker;
 import mods.belgabor.amtweaker.mods.amt.handlers.Plate;
 import mods.belgabor.amtweaker.mods.amt.handlers.Processor;
 
@@ -10,5 +11,6 @@ public class AMT {
         MineTweakerAPI.registerClass(Evaporator.class);
         MineTweakerAPI.registerClass(Plate.class);
         MineTweakerAPI.registerClass(Processor.class);
+        MineTweakerAPI.registerClass(IceMaker.class);
     }
 }