+++ changelog.md
@@ -2,10 +2,11 @@ Changelog
 =========
 
 Version 0.1alpha
------------
+----------------
 Early alpha version
 
 * Cooking Iron Plate (add/remove recipe, add heat source)
 * Evaporator (add/remove recipe)
 * Food Processor/Jaw Crusher (add/remove recipe)
+* Ice Maker (add/remove recipe, register/unregister charge item)